Studio: make plugin registry lookups independent of plugin init order
Resolving a service another plugin binds in its own onInit depends on plugin
init order, which is the module federation remote load order whenever the
priorities are equal. That made the IndexBundle plugin fail with

  No matching bindings found for serviceIdentifier:
  CoreShop/StudioForm/WidgetRegistry

and, because the block is wrapped in try/catch for the reduced document editor
iframe, the failure was swallowed together with the pre/user condition registry
binds that followed it.

Two changes:

- Give the provider plugins a negative priority so their binds always run
  first: coreshop-pimcore (-1200), coreshop-resource (-1100),
  coreshop-studio-form-plugin (-1000, already present). Plugins are sorted
  ascending by priority with a default of 0, so consumers keep running after
  them without any change of their own.

- Move IndexBundle's widget, object data and schema widget registrations out of
  onInit into FiltersModule and IndexesModule, registered from onStartup.
  Modules are initialised after every plugin's onInit, so cross-plugin
  registries are guaranteed to exist. Only this bundle's own container binds
  stay in onInit, where they cannot fail.

Fixes #3141
